Added Fosh.html V1
First edition of the bot; incapable of piece value recognition, or a basic opening strategy. This version of Fosh uses extensive JavaScript in an HMTL file, which may be blocked by many browsers. I plan to create a version that relies on an external Foss.js file to handle the script. Currently, it has about 1008 lines, mainly composed of, you guessed it, JavaScript.